# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Prueba API",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://swagger.local"
        }
    ],
    "paths": {
...
        "/api/loginUser": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Login Usuario.",
                "operationId": "App\\Http\\Controllers\\UserController::authenticate",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "email": " ",
                                    "password": " "
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login de usuario ."
                    },
                    "500": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
        "/api/registrarUser": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Registrar Usuario.",
                "operationId": "App\\Http\\Controllers\\UserController::registrarUser",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "nombre",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "apellido",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "fecha_nacimiento",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "date"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "nombre": {
                                        "type": "string"
                                    },
                                    "apellido": {
                                        "type": "string"
                                    },
                                    "fecha_nacimiento": {
                                        "type": "date"
                                    },
                                    "password": {
                                        "type": "date"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "email": "admin@gmail.com",
                                    "nombre": "Jose",
                                    "apellido": "Regalado",
                                    "fecha_nacimiento": "2020-10-10",
                                    "password": "000000"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Registrar Usuario ."
                    },
                    "500": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
        "/api/editarUser": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Editar usuario",
                "operationId": "App\\Http\\Controllers\\UserController::editarUser",
                "parameters": [
                    {
                        "name": "nombre",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "apellido",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "fecha_nacimiento",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "date"
                        }
                    },
                    {
                        "name": "activo",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombre": {
                                        "type": "string"
                                    },
                                    "apellido": {
                                        "type": "string"
                                    },
                                    "fecha_nacimiento": {
                                        "type": "date"
                                    },
                                    "activo": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "nombre": "Jose",
                                    "apellido": "Regalado",
                                    "fecha_nacimiento": "2020-10-10",
                                    "activo": true
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Editar Usuario ."
                    },
                    "500": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        }
    },
    "security": [
        []
    ]
}